Create Dawg.py
This bot was custom-made to interact with a specific user that is not me. Because of this, I have changed some lines to protect their personal information. (Hint: I replaced their Twitter handle with "account" on line 21, their nickname with just "Dawg" on multiple lines, and two pieces of text in lines 36 and 37 used to ignore tweets with specific key words).